Base64Value moved to the common/common module
