+++ common/crypto/src/main/java/io/helidon/common/crypto/AsymmetricCipher.java
@@ -23,7 +23,7 @@ import java.util.Objects;
 
 import javax.crypto.Cipher;
 
-import io.helidon.integrations.common.rest.Base64Value;
+import io.helidon.common.Base64Value;
 
 /**
  * This class provides simple and stateless way to encrypt and decrypt messages using selected asymmetric cipher.
