+++ common/crypto/src/main/java/io/helidon/common/crypto/Digest.java
@@ -17,10 +17,9 @@
 package io.helidon.common.crypto;
 
 import java.security.MessageDigest;
-import java.util.Arrays;
 import java.util.regex.Matcher;
 
-import io.helidon.integrations.common.rest.Base64Value;
+import io.helidon.common.Base64Value;
 
 import static io.helidon.common.crypto.CryptoCommonConstants.PREFIX;
 import static io.helidon.common.crypto.CryptoCommonConstants.PREFIX_PATTERN;
