+++ examples/integrations/oci/vault-reactive/src/main/java/io/helidon/examples/integrations/oci/vault/reactive/VaultService.java
@@ -21,7 +21,7 @@ import java.time.temporal.ChronoUnit;
 import java.util.Optional;
 
 import io.helidon.common.http.Http;
-import io.helidon.integrations.common.rest.Base64Value;
+import io.helidon.common.Base64Value;
 import io.helidon.integrations.oci.vault.CreateSecret;
 import io.helidon.integrations.oci.vault.Decrypt;
 import io.helidon.integrations.oci.vault.DeleteSecret;
