+++ integrations/oci/vault/src/main/java/io/helidon/integrations/oci/vault/Decrypt.java
@@ -21,7 +21,7 @@ import java.util.Optional;
 import javax.json.JsonObject;
 
 import io.helidon.integrations.common.rest.ApiEntityResponse;
-import io.helidon.integrations.common.rest.Base64Value;
+import io.helidon.common.Base64Value;
 import io.helidon.integrations.oci.connect.OciApiException;
 import io.helidon.integrations.oci.connect.OciRequestBase;
