+++ integrations/oci/vault/src/main/java/io/helidon/integrations/oci/vault/Encrypt.java
@@ -21,7 +21,7 @@ import java.util.Optional;
 import javax.json.JsonObject;
 
 import io.helidon.integrations.common.rest.ApiEntityResponse;
-import io.helidon.integrations.common.rest.Base64Value;
+import io.helidon.common.Base64Value;
 import io.helidon.integrations.oci.connect.OciApiException;
 import io.helidon.integrations.oci.connect.OciRequestBase;
 
@@ -71,8 +71,8 @@ public final class Encrypt {
          *
          * @param value value to encrypt
          * @return updated request
-         * @see io.helidon.integrations.common.rest.Base64Value#create(String)
-         * @see io.helidon.integrations.common.rest.Base64Value#create(byte[])
+         * @see Base64Value#create(String)
+         * @see Base64Value#create(byte[])
          */
         public Request data(Base64Value value) {
             return add("plaintext", value.toBase64());
