+++ integrations/oci/vault/src/main/java/io/helidon/integrations/oci/vault/OciVaultSecurityProvider.java
@@ -24,7 +24,7 @@ import java.util.function.Supplier;
 import io.helidon.common.reactive.Single;
 import io.helidon.config.Config;
 import io.helidon.integrations.common.rest.ApiOptionalResponse;
-import io.helidon.integrations.common.rest.Base64Value;
+import io.helidon.common.Base64Value;
 import io.helidon.security.spi.DigestProvider;
 import io.helidon.security.spi.EncryptionProvider;
 import io.helidon.security.spi.ProviderConfig;
