+++ integrations/vault/secrets/transit/src/main/java/io/helidon/integrations/vault/secrets/transit/DecryptBatch.java
@@ -26,7 +26,7 @@ import javax.json.JsonValue;
 import io.helidon.integrations.common.rest.ApiEntityResponse;
 import io.helidon.integrations.common.rest.ApiException;
 import io.helidon.integrations.common.rest.ApiJsonBuilder;
-import io.helidon.integrations.common.rest.Base64Value;
+import io.helidon.common.Base64Value;
 import io.helidon.integrations.vault.VaultRequest;
 import io.helidon.integrations.vault.VaultResponse;
