+++ integrations/vault/secrets/transit/src/main/java/io/helidon/integrations/vault/secrets/transit/Sign.java
@@ -20,7 +20,7 @@ import javax.json.JsonObject;
 
 import io.helidon.integrations.common.rest.ApiEntityResponse;
 import io.helidon.integrations.common.rest.ApiException;
-import io.helidon.integrations.common.rest.Base64Value;
+import io.helidon.common.Base64Value;
 import io.helidon.integrations.vault.VaultRequest;
 import io.helidon.integrations.vault.VaultResponse;
 
@@ -132,8 +132,8 @@ public final class Sign {
          *
          * @param value value to encrypt
          * @return updated request
-         * @see io.helidon.integrations.common.rest.Base64Value#create(String)
-         * @see io.helidon.integrations.common.rest.Base64Value#create(byte[])
+         * @see Base64Value#create(String)
+         * @see Base64Value#create(byte[])
          */
         public Request data(Base64Value value) {
             return add("input", value.toBase64());
